raft全部使用64位整形

+++ src/cr/raft/candidate.cpp
@@ -169,7 +169,7 @@ namespace cr
             return false;
         }
 
-        void Candidate::setNewerTerm(std::uint32_t newerTerm)
+        void Candidate::setNewerTerm(std::uint64_t newerTerm)
         {
             engine.setCurrentTerm(newerTerm);
             engine.setVotedFor(boost::none);
